requirements.txt
fastapi
uvicorn
pydantic
pytest
pillow
python-multipart
httpx